IGAPP-531: remove useEffect and pass defaultViewport calculated by BBox
